+++ src/helperFunctions.ts
@@ -1,0 +1,5 @@
+
+export function getCurrentDateTime() {
+    return new Date().toISOString();
+  }
+
